# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  gin.H:
    additionalProperties: {}
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@omninudge.com
    name: OmniNudge Support
  description: OmniNudge social platform API
  license:
    name: Proprietary
  termsOfService: http://omninudge.com/terms/
  title: OmniNudge API
  version: "1.0"
paths:
...
  /hubs/{name}/theme:
    get:
      parameters:
      - description: Hub name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/gin.H'
      summary: Get active hub theme
      tags:
      - Hubs
  /hubs/{name}/themes:
    get:
      parameters:
      - description: Hub name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Get all hub themes
      tags:
      - Hubs
    post:
      consumes:
      - application/json
      parameters:
      - description: Hub name
        in: path
        name: name
        required: true
        type: string
      - description: Theme payload
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Create hub theme
      tags:
      - Hubs
  /hubs/{name}/themes/{id}:
    delete:
      parameters:
      - description: Hub name
        in: path
        name: name
        required: true
        type: string
      - description: Theme ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Delete hub theme
      tags:
      - Hubs
    put:
      consumes:
      - application/json
      parameters:
      - description: Hub name
        in: path
        name: name
        required: true
        type: string
      - description: Theme ID
        in: path
        name: id
        required: true
        type: integer
      - description: Theme payload
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Update hub theme
      tags:
      - Hubs
  /hubs/{name}/themes/{id}/activate:
    post:
      parameters:
      - description: Hub name
        in: path
        name: name
        required: true
        type: string
      - description: Theme ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Activate hub theme
      tags:
      - Hubs
  /hubs/{name}/themes/preview:
    post:
      consumes:
      - application/json
      parameters:
      - description: Hub name
        in: path
        name: name
        required: true
        type: string
      - description: Theme CSS payload
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Preview hub theme CSS
      tags:
      - Hubs
...
securityDefinitions:
  BearerAuth:
    description: 'JWT Bearer token. Format: "Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"